refactor(agents): split chat area components
Extract message rendering, tool output, empty state, and header controls.

Keep AgentsChatArea focused on state and chat transport orchestration.
